preentrega 3 casi finalizada, falta anteultimo punto

+++ src/routes/products.router.js
@@ -3,7 +3,8 @@ import { productsModel } from "../dao/models/products.model.js";
 /* import { ProductManager } from '../dao/fileManagers/ProductManager.js'; */
 import ProductManager from '../dao/mongoManagers/ProductManager.js'
 import { faker } from '@faker-js/faker';
-
+import { jwtValidation } from "../middlewares/jwt.middleware.js";
+import mongoose from "mongoose";
 
 const router = Router()
 
@@ -57,9 +58,30 @@ router.get('/:idProducts', async (req, res) => {
 
 
 /* Add product */
-router.post('/', async (req, res) => {
+/* router.post('/', async (req, res) => {
     try {
         const newProduct = await pm.addProduct(req.body);
+        res.json({ message: 'Producto creado con éxito', newProduct });
+    } catch (error) {
+        console.error(error);
+        res.status(500).json({ message: 'Error al crear el producto' });
+    }
+}); */
+
+router.post('/', async (req, res) => {
+    try {
+        const user = req.user; // Obtener el usuario actual
+
+        // Verificar el rol del usuario y asignar el propietario en función de ello
+        let owner = "admin";
+        if (user.role === "premium") {
+            owner = "premium";
+        }
+
+        let createdBy = user._id
+
+        // Crear el nuevo producto con el propietario asignado
+        const newProduct = await pm.addProduct({ ...req.body, owner, createdBy });
         res.json({ message: 'Producto creado con éxito', newProduct });
     } catch (error) {
         console.error(error);
@@ -69,8 +91,9 @@ router.post('/', async (req, res) => {
 
 
 
+
 /* Modify product */
-router.post('/:idProducts', async (req, res) => {
+/*  router.post('/:idProducts', async (req, res) => {
     const { idProducts } = req.params;
     const updateFields = req.body;
 
@@ -90,6 +113,35 @@ router.post('/:idProducts', async (req, res) => {
         console.error(error);
         res.status(500).json({ message: "Error al editar el producto" });
     }
+});  */
+
+router.post('/:idProducts', jwtValidation, async (req, res) => {
+    const { idProducts } = req.params;
+    const updateFields = req.body;
+    const userId = req.user._id;
+    const isAdmin = req.user.role === 'admin';
+
+    try {
+        const product = await productsModel.findById(idProducts);
+        if (!product) {
+            return res.status(404).json({ message: "Producto no encontrado" });
+        }
+
+        // Verificar si el producto pertenece al usuario o si el usuario es un administrador
+        if (isAdmin || (product.owner === userId)) {
+            for (const [key, value] of Object.entries(updateFields)) {
+                product[key] = value;
+            }
+
+            const updatedProduct = await product.save();
+            res.json({ message: "Producto editado correctamente", product: updatedProduct });
+        } else {
+            res.status(403).send('No tienes permiso para modificar este producto');
+        }
+    } catch (error) {
+        console.error(error);
+        res.status(500).json({ message: "Error al editar el producto" });
+    }
 });
 
 
@@ -98,20 +150,34 @@ router.post('/:idProducts', async (req, res) => {
 
 router.post('/delete/:id', async (req, res) => {
     const { id } = req.params;
+    const user = req.user; // Obtener el usuario actual
 
     try {
-        const deletedProduct = await productsModel.findByIdAndDelete(id);
-        if (deletedProduct) {
-            res.json({ message: `Producto con ID ${id} eliminado correctamente.` });
-        } else {
-            res.status(404).json({ message: `No se encontró el producto con ID ${id}.` });
+        const product = await productsModel.findById(id);
+
+        if (!product) {
+            return res.status(404).json({ message: `No se encontró el producto con ID ${id}.` });
         }
+
+        if (user.role === 'admin') {
+            // Si el usuario es un administrador, puede eliminar cualquier producto
+            await productsModel.findByIdAndDelete(id);
+            return res.json({ message: `Producto con ID ${id} eliminado correctamente.` });
+        }
+
+        if (user.role === 'premium' && product.owner === 'premium') {
+            // Si el usuario es premium y el producto le pertenece, puede eliminarlo
+            await productsModel.findByIdAndDelete(id);
+            return res.json({ message: `Producto con ID ${id} eliminado correctamente.` });
+        }
+
+        // Si el usuario premium intenta eliminar un producto que no le pertenece, se devuelve un error
+        return res.status(403).json({ message: 'No tienes permiso para eliminar este producto' });
     } catch (error) {
         console.error(error);
-        res.status(500).json({ message: "Error al eliminar el producto" });
+        res.status(500).json({ message: 'Error al eliminar el producto' });
     }
 });
-
